+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mkhuda/react-shaka-player",
-  "version": "1.1.5",
+  "version": "1.1.6-beta.0",
   "description": "React video player built with Shaka-Player",
   "main": "dist/index.js",
   "author": "Muhammad K. Huda",
